Add files via upload
Easy to guess passwords often times are not strong enough to secure online accounts or any account for that matter. With this code, you can generate a very strong password that is difficult to crack. Each time the code runs, its generates a new sequence of 10 character passwords made up of upper case alphabets, lowercase alphabets, numbers 1 to 10 and special symbols.